backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
requests==2.32.3
django-filter==25.1
celery==5.5.1
redis==5.3.0
